Isolate test fixtures for service tests (#1064)
Now test classes that inherit from WebServiceTestCase all need to
explicitly provision the documents that they need to be indexed, instead
of just always indexing a document in WebServiceTestCase.
