+++ test/service/test_dss_proxy.py
@@ -1,5 +1,4 @@
 import io
-import json
 import logging
 import os
 import time
